+++ CHANGELOG.rst
@@ -11,6 +11,8 @@ This project adheres to `Semantic Versioning`_ starting with version 0.2.0.
 
 Added
 -----
+- new ``RasaChatInput`` channel
+- new ``NoOutputChannel`` which does nothing
 
 Changed
 -------